chore(db): fold mutators and the write queue into the write compartment
