+++ crates/db/src/write/autocommit/helpers.rs
