+++ crates/db/src/write/mod.rs
@@ -1,0 +1,4 @@
+//! The write path: the mutators that produce document deltas.
+pub mod autocommit;
+pub(crate) mod doc;
+pub mod queue;
