+++ proofs/tla/InteractiveTxnCounter.tla
@@ -1,7 +1,7 @@
 ---- MODULE InteractiveTxnCounter ----
 \* Guard-acquisition LIFECYCLE of the interactive / explicit-transaction counter mutator,
-\* abstracting crates/db/src/doc_mutator.rs (DbDocMutator), crates/db/src/txn/mod.rs (DbTxn
-\* pending counter deltas + commit) and crates/db/src/doc_write_queue.rs (acquire /
+\* abstracting crates/db/src/write/doc.rs (DbDocMutator), crates/db/src/txn/mod.rs (DbTxn
+\* pending counter deltas + commit) and crates/db/src/write/queue.rs (acquire /
 \* acquire_batch_gate). This is the #1044 design, IMPLEMENTED in the tree (commit-time
 \* finalize in txn/registry/lifecycle.rs::finalize_and_commit / apply_counter_ops_at_finalize — see
 \* InteractiveTxnCounter_DESIGN.md). Companion to MergeQueue.tla, which models the per-doc
